+++ betfairlightweight/endpoints/streaming.py
@@ -56,7 +56,6 @@ class Streaming:
         file_path: str = None,
         listener: BaseListener = None,
         operation: str = "marketSubscription",
-        **kwargs
     ) -> HistoricalStream:
         """
         Uses streaming listener/cache to parse betfair
@@ -69,22 +68,14 @@ class Streaming:
 
         :rtype: HistoricalStream
         """
-        if file_path is None and kwargs.get("directory"):
-            warnings.warn(
-                "directory is deprecated; use file_path", DeprecationWarning,
-            )
-            file_path = kwargs.get("directory")
-
         listener = listener if listener else BaseListener()
-        listener.register_stream(0, operation)
-        return HistoricalStream(file_path, listener)
+        return HistoricalStream(file_path, listener, operation)
 
     @staticmethod
     def create_historical_generator_stream(
         file_path: str = None,
         listener: BaseListener = None,
         operation: str = "marketSubscription",
-        **kwargs
     ) -> HistoricalGeneratorStream:
         """
         Uses generator listener/cache to parse betfair
@@ -97,12 +88,5 @@ class Streaming:
 
         :rtype: HistoricalGeneratorStream
         """
-        if file_path is None and kwargs.get("directory"):
-            warnings.warn(
-                "directory is deprecated; use file_path", DeprecationWarning,
-            )
-            file_path = kwargs.get("directory")
-
         listener = listener if listener else StreamListener()
-        listener.register_stream(0, operation)
-        return HistoricalGeneratorStream(file_path, listener)
+        return HistoricalGeneratorStream(file_path, listener, operation)