+++ tests/unit/test_betfairstream.py
@@ -357,7 +357,9 @@ class HistoricalStreamTest(unittest.TestCase):
     def setUp(self):
         self.file_path = "tests/resources/historicaldata/BASIC-1.132153978"
         self.listener = mock.Mock()
-        self.stream = HistoricalStream(self.file_path, self.listener)
+        self.stream = HistoricalStream(
+            self.file_path, self.listener, "marketSubscription"
+        )
 
     def test_init(self):
         assert self.stream.file_path == self.file_path
@@ -383,13 +385,16 @@ class HistoricalStreamTest(unittest.TestCase):
         self.listener.on_data.snap()
         mock_stop.assert_called_with()
         self.assertTrue(self.stream._running)
+        self.listener.register_stream.assert_called_with(0, "marketSubscription")
 
 
 class HistoricalGeneratorStreamTest(unittest.TestCase):
     def setUp(self):
         self.file_path = "tests/resources/historicaldata/BASIC-1.132153978"
         self.listener = mock.Mock()
-        self.stream = HistoricalGeneratorStream(self.file_path, self.listener)
+        self.stream = HistoricalGeneratorStream(
+            self.file_path, self.listener, "marketSubscription"
+        )
 
     def test_init(self):
         assert self.stream.file_path == self.file_path
@@ -412,3 +417,4 @@ class HistoricalGeneratorStreamTest(unittest.TestCase):
         self.listener.on_data.snap()
         mock_stop.assert_called_with()
         self.assertTrue(self.stream._running)
+        self.listener.register_stream.assert_called_with(0, "marketSubscription")